mcp-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "modexia-mcp"
version = "0.1.0"
description = "Official Modexia AgentPay MCP Server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.26.0",
    "modexiaagentpay>=0.5.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]

[project.scripts]
modexia-mcp = "server:main"
